requirements.txt
apache-airflow==3.0.6
apache-airflow-providers-google==17.1.0
apache-airflow-providers-postgres==6.2.3
ruff==0.12.11
